--- packages/core/src/closure.ts
@@ -1,10 +1,0 @@
-export function useState<T>(initialValue: T) {
-  let value = initialValue
-  function get(): T {
-    return value
-  }
-  function set(newVal: T): void {
-    value = newVal
-  }
-  return { get, set }
-}
